requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
pytest
allure-pytest
